# [NOSQUASH] Precompiled headers support