Recent CMake has a `target_precompile_headers` function that makes using precompiled headers pretty trivial.
(For comparison, see #8643. (cc @numberZero))

Goal: Faster compilation, climate protection, less turbine noise.

1st half of commit:
Moves sha256.c to lib/. This is needed because precompiled headers don't work (assume C) if we're mixing C and C++.

2nd half of commit:
Precompiled header option. (Requires new enough cmake.)

Which headers are precompiled:
* We only want to precompile headers that seldom or never change. Otherwise we'd have to recompile the precompiled header and recompile everything else when the header changes. And ccache can't know if the precompiled header is different, so even just switching between branches would be horribly painful.
* => Only stdlib and lib headers are precompiled (excluding irrlicht).
* I've taken the sdlib headers from #8643, and added jsoncpp. (Could add more.)
* The list of headers is read from a file (`src/precompiled_headers.txt`). The user can configure this.

## To do

This PR is a Ready for Review.

## How to test

* If you're using ccache, add to your ccache.conf: `sloppiness = pch_defines,time_macros`
* `$ cmake . -DPRECOMPILE_HEADERS=1`
* `$ make -j<you choose>`

Measure speed:
* `$ cmake . -DCMAKE_CXX_COMPILER=/usr/bin/g++`
* `$ make -B -j5`
* Results: (may be inaccurate)
  * No precompile:
    ```
    real	7m31,836s
    user	34m38,999s
    sys	1m53,541s
    ```
  * Precompile:
    ```
    real	6m16,476s
    user	28m39,059s
    sys	1m32,657s
    ```
  * => ca. 17% faster