@lhofhansl 2023-03-26
I'd be worried the additional fragility in the process is not warranted by a 17% compilation time decrease.