@numberZero 2023-03-26
I tested this (8812b793d) with 3 build directories (warm-up, PCH disabled, and PCH enabled) on my system (Ryzen 7 5800X). Here is time comparison:

Target | Real, no PCH | Real, PCH | Real, diff | CPU, no PCH | CPU, PCH | CPU, diff
----|---|---|---|---|---|---
Clean build | 56 | 48 | −15% | 785 | 629 | −20%
Rebuild after touching client/content_mapblock.cpp | 6.0 | 5.6 | −6% | 6.0 | 5.6 | −6%
Rebuild after touching client/tile.h | 41 | 35 | −16% | 542 | 437 | −19%

Time is in seconds. Time diff is percentage of “no PCH.”

> 17% faster

Corresponds to −17% “Real, diff” (Clean build) in my format. Which is just a bit better than what I get.

Here are some details:

### Without PCH
```
$ cmake -GNinja -DCMAKE_BUILD_TYPE=Debug -DRUN_IN_PLACE=ON -DPRECOMPILE_HEADERS=OFF
<...>
[482/482] Linking CXX executable <...>/bin/minetest
real    0m56,241s
user    12m13,479s
sys     0m51,037s

$ touch src/client/content_mapblock.cpp
<...>
[4/4] Linking CXX executable <...>/bin/minetest
real    0m5,991s
user    0m5,480s
sys     0m0,510s

$ touch src/client/tile.h
<...>
[178/178] Linking CXX executable <...>/bin/minetest
real    0m41,414s
user    8m29,035s
sys     0m32,512s
```

### With PCH
```
$ cmake -GNinja -DCMAKE_BUILD_TYPE=Debug -DRUN_IN_PLACE=ON -DPRECOMPILE_HEADERS=ON
<...>
[483/483] Linking CXX executable <...>/bin/minetest
real    0m48,005s
user    9m41,919s
sys     0m46,652s

$ touch src/client/content_mapblock.cpp
<...>
[6/6] Linking CXX executable <...>/bin/minetest
real    0m5,616s
user    0m5,125s
sys     0m0,508s

$ touch src/client/tile.h
<...>
[177/177] Linking CXX executable <...>/bin/minetest
real    0m34,724s
user    6m47,845s
sys     0m29,482s
```

### Conclusion

While single source file updates are dominated by link time, there is a clear speedup for *header* updates which tend to require massive recompilation.