@Desour 2023-03-26
> I'd be worried the additional fragility in the process is not warranted by a 17% compilation time decrease.

It's optional and marked as experimental. And CI doesn't use it, so missing includes and similar should usually be found.

Thanks for testing, @numberZero!