@numberZero 2023-03-26
> missing includes

Which may very well happen without any PCH because e.g. today, `<vector>` may include `<cstdint>` so your code will compile fine... until STL upgrade.